requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiogram-dialog==2.0.0b18
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
build==0.10.0
CacheControl==0.12.11
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==40.0.2
distlib==0.3.6
dulwich==0.21.3
filelock==3.12.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
magic-filter==1.0.9
MarkupSafe==2.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
oauthlib==3.2.2
packaging==23.1
pendulum==2.1.2
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.1
protobuf==4.22.3
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
rapidfuzz==2.15.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.7
trove-classifiers==2023.4.22
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.1
webencodings==0.5.1
yarl==1.9.1
zipp==3.15.0
